# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BacktestPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BacktestResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before"
          }
        },
        "required": [
          "items",
          "next_before"
        ],
        "title": "BacktestPage",
        "type": "object"
      },
...
      "BacktestRequest": {
        "description": "`strategy_id` comes from the path \u2014 it is not part of the body.",
        "properties": {
          "commission_bps": {
            "default": 1.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Commission Bps",
            "type": "number"
          },
          "initial_capital": {
            "default": 10000.0,
            "exclusiveMinimum": 0.0,
            "title": "Initial Capital",
            "type": "number"
          },
          "lookback_minutes": {
            "default": 240,
            "maximum": 1440.0,
            "minimum": 5.0,
            "title": "Lookback Minutes",
            "type": "integer"
          },
          "slippage_bps": {
            "default": 0.5,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Slippage Bps",
            "type": "number"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "title": "BacktestRequest",
        "type": "object"
      },
      "BacktestResponse": {
        "properties": {
          "commission_bps": {
            "title": "Commission Bps",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data_source": {
            "title": "Data Source",
            "type": "string"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "exposure_pct": {
            "title": "Exposure Pct",
            "type": "number"
          },
          "final_capital": {
            "title": "Final Capital",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "initial_capital": {
            "title": "Initial Capital",
            "type": "number"
          },
          "max_drawdown_pct": {
            "title": "Max Drawdown Pct",
            "type": "number"
          },
          "profit_factor": {
            "title": "Profit Factor",
            "type": "number"
          },
          "sharpe_ratio": {
            "title": "Sharpe Ratio",
            "type": "number"
          },
          "slippage_bps": {
            "title": "Slippage Bps",
            "type": "number"
          },
          "sortino_ratio": {
            "title": "Sortino Ratio",
            "type": "number"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "strategy_id": {
            "title": "Strategy Id",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "total_pnl": {
            "title": "Total Pnl",
            "type": "number"
          },
          "total_pnl_pct": {
            "title": "Total Pnl Pct",
            "type": "number"
          },
          "total_trades": {
            "title": "Total Trades",
            "type": "integer"
          },
          "win_rate": {
            "title": "Win Rate",
            "type": "number"
          }
        },
        "required": [
          "id",
          "strategy_id",
          "symbol",
          "data_source",
          "status",
          "initial_capital",
          "final_capital",
          "total_pnl",
          "total_pnl_pct",
          "total_trades",
          "win_rate",
          "max_drawdown_pct",
          "sharpe_ratio",
          "sortino_ratio",
          "profit_factor",
          "exposure_pct",
          "commission_bps",
          "slippage_bps",
          "start_time",
          "end_time",
          "created_at"
        ],
        "title": "BacktestResponse",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "StrategyCreate": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "strategy_type": {
            "default": "rule_based",
            "enum": [
              "rule_based",
              "ml_signal",
              "hybrid"
            ],
            "title": "Strategy Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "config"
        ],
        "title": "StrategyCreate",
        "type": "object"
      },
      "StrategyExample": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "description": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Description",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "config",
          "description"
        ],
        "title": "StrategyExample",
        "type": "object"
      },
      "StrategyResponse": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_public": {
            "title": "Is Public",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "strategy_type": {
            "title": "Strategy Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "description",
          "strategy_type",
          "config",
          "is_active",
          "is_public",
          "created_at",
          "updated_at"
        ],
        "title": "StrategyResponse",
        "type": "object"
      },
      "StrategyUpdate": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "strategy_type": {
            "anyOf": [
              {
                "enum": [
                  "rule_based",
                  "ml_signal",
                  "hybrid"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Type"
          }
        },
        "title": "StrategyUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Real-time market-microstructure intelligence: streaming features, order-book analytics, calibrated ML signals, alerts and backtesting.",
    "title": "AlgoViz",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/strategies": {
      "get": {
        "operationId": "list_strategies_api_v1_strategies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrategyResponse"
                  },
                  "title": "Response List Strategies Api V1 Strategies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "List strategies",
        "tags": [
          "Strategies"
        ]
      },
      "post": {
        "operationId": "create_strategy_api_v1_strategies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create strategy",
        "tags": [
          "Strategies"
        ]
      }
    },
    "/api/v1/strategies/examples": {
      "get": {
        "operationId": "list_examples_api_v1_strategies_examples_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrategyExample"
                  },
                  "title": "Response List Examples Api V1 Strategies Examples Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Example strategy specs",
        "tags": [
          "Strategies"
        ]
      }
    },
    "/api/v1/strategies/{strategy_id}": {
      "delete": {
        "operationId": "delete_strategy_api_v1_strategies__strategy_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete strategy",
        "tags": [
          "Strategies"
        ]
      },
      "get": {
        "operationId": "get_strategy_api_v1_strategies__strategy_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Get strategy",
        "tags": [
          "Strategies"
        ]
      },
      "patch": {
        "operationId": "update_strategy_api_v1_strategies__strategy_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update strategy",
        "tags": [
          "Strategies"
        ]
      }
    },
    "/api/v1/strategies/{strategy_id}/backtest": {
      "post": {
        "operationId": "run_backtest_api_v1_strategies__strategy_id__backtest_post",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too many backtests running or queued"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Run a backtest (async; progress streams on the `backtests` WS channel)",
        "tags": [
          "Strategies"
        ]
      }
    },
    "/api/v1/strategies/{strategy_id}/backtests": {
      "get": {
        "operationId": "list_backtests_api_v1_strategies__strategy_id__backtests_get",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "`next_before` of the previous page",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`next_before` of the previous page",
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "List backtests, newest first",
        "tags": [
          "Strategies"
        ]
      }
    },
    "/api/v1/strategies/{strategy_id}/backtests/{backtest_id}": {
      "get": {
        "operationId": "get_backtest_api_v1_strategies__strategy_id__backtests__backtest_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "backtest_id",
            "required": true,
            "schema": {
              "title": "Backtest Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Backtest detail with trades and equity curve",
        "tags": [
          "Strategies"
        ]
      }
    },
...
    "/api/v1/ws/schema": {
      "get": {
        "description": "Exists so the server\u2192client and client\u2192server WebSocket message unions are part of the OpenAPI document (and therefore the generated TypeScript types). Always 204.",
        "operationId": "ws_schema_api_v1_ws_schema_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WSSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "No content \u2014 see components.schemas.WSSchema"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "WebSocket message schemas (OpenAPI anchor; responds 204)",
        "tags": [
          "System"
        ]
      }
    },
...
  }
}